Single block map mesh generation of 3D volume
An arbitrary 3D faceted domain with six connected faces is selected. The domain is decomposed into separate faces consists of triangular faces. This was done by creating half edge structure of the connected triangles and find out 6 discreate faces which are separated by a given specified feature angle. The connected edges of the individual faces are identified and those edges are parametrized with number of input divisions. Then those boundary values are interpolated by transfinite interpolation to create the 3d structured mesh.